Switching to generalized cubic spline generators. Untested for now! Doing that next
