+++ solvers/qpSpline/nWaypointCubicSplinemex.cpp
@@ -8,11 +8,11 @@ using namespace std;
 using namespace Eigen;
 
 void mexFunction(int nlhs, mxArray *plhs[],int nrhs, const mxArray *prhs[]) {
-  string usage = "[coefs, objval] = twoWaypointCubicSplinemex(ts, xs, xd0, xdf)";
+  string usage = "[coefs, objval] = nWaypointCubicSplinemex(ts, xs, xd0, xdf)";
   if (nrhs != 4)
-    mexErrMsgIdAndTxt("Drake:twoWaypointCubicSplinemex:WrongNumberOfInputs", usage.c_str());
+    mexErrMsgIdAndTxt("Drake:nWaypointCubicSplinemex:WrongNumberOfInputs", usage.c_str());
   if (nlhs > 2)
-    mexErrMsgIdAndTxt("Drake:twoWaypointCubicSplinemex:WrongNumberOfOutputs", usage.c_str());
+    mexErrMsgIdAndTxt("Drake:nWaypointCubicSplinemex:WrongNumberOfOutputs", usage.c_str());
 
   const std::vector<double> segment_times = matlabToStdVector(prhs[0]);
   MatrixXd xs = matlabToEigen<Dynamic, Dynamic>(prhs[1]);
@@ -20,13 +20,18 @@ void mexFunction(int nlhs, mxArray *plhs[],int nrhs, const mxArray *prhs[]) {
   auto xdf = matlabToEigen<Dynamic, 1>(prhs[3]);
 
   mwSize ndof = static_cast<mwSize>(xs.rows());
-  mwSize num_segments = 3;
+  mwSize num_segments = static_cast<mwSize>(xs.cols())-1;
+  mwSize num_knots = num_segments - 1;
   mwSize num_coeffs_per_segment = 4;
   mwSize dims[] = {ndof, num_segments, num_coeffs_per_segment};
   plhs[0] = mxCreateNumericArray(num_segments, dims, mxDOUBLE_CLASS, mxREAL);
   double objective_value = 0.0;
   for (mwSize dof = 0; dof < ndof; dof++) {
-    PiecewisePolynomial spline = twoWaypointCubicSpline(segment_times, xs(dof, 0), xd0[dof], xs(dof, 3), xdf[dof], xs(dof, 1), xs(dof, 2));
+    std::vector<double> xi(num_knots);
+    for (int knot = 0; knot < num_knots; knot++)
+      xi[knot] = xs(dof, knot);
+
+    PiecewisePolynomial spline = nWaypointCubicSpline(segment_times, xs(dof, 0), xd0[dof], xs(dof, num_segments), xdf[dof], xi);
 
     PiecewisePolynomial acceleration_squared = spline.derivative(2);
     acceleration_squared *= acceleration_squared;
@@ -36,7 +41,7 @@ void mexFunction(int nlhs, mxArray *plhs[],int nrhs, const mxArray *prhs[]) {
     for (mwSize segment_index = 0; segment_index < spline.getNumberOfSegments(); segment_index++) {
       for (mwSize coefficient_index = 0; coefficient_index < num_coeffs_per_segment; coefficient_index++) {
         mwSize sub[] = {dof, segment_index, num_coeffs_per_segment - coefficient_index - 1}; // Matlab's reverse coefficient indexing...
-        *(mxGetPr(plhs[0]) + sub2ind(3, dims, sub)) = spline.getPolynomial(segment_index).getCoefficients()[coefficient_index];
+        *(mxGetPr(plhs[0]) + sub2ind(num_segments, dims, sub)) = spline.getPolynomial(segment_index).getCoefficients()[coefficient_index];
       }
     }
   }
